backend/requirements.txt
# AutoOps Backend Dependencies (Retool Integration)

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# HTTP Client (for calling Retool webhook)
httpx==0.28.0

# Data Validation
pydantic==2.10.0

# Configuration
python-dotenv==1.0.0
